requirements.txt
# Core Flask Dependencies
Flask>=3.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
Flask-WTF>=1.0.0
Flask-Mail>=0.9.0
Jinja2>=3.0.0
Werkzeug>=2.0.0
WTForms>=3.0.0

# Database Dependencies
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0

# Web Server
gunicorn>=20.0.0

# Authentication & Security
PyJWT>=2.8.0
email-validator>=2.0.0

# Environment & Configuration
python-dotenv>=1.0.0

# Payment Processing
stripe>=6.0.0

# Timezone & Date Handling
pytz>=2023.0
python-dateutil>=2.8.0

# HTTP Requests
requests>=2.25.0

# Scheduling System Dependencies
schedule>=1.0.0

# Video functionality has been removed from this application

# Additional Utilities
click>=8.0.0
MarkupSafe>=2.0.0 